+++ publish-service/src/main/java/com/ericsson/eiffel/remrem/publish/controller/ProducerController.java
@@ -1,5 +1,5 @@
 /*
-    Copyright 2017 Ericsson AB.
+    Copyright 2018 Ericsson AB.
     For a full list of individual contributors, please see the commit history.
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -19,7 +19,6 @@ import java.util.Map;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -34,56 +33,71 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
 import com.ericsson.eiffel.remrem.protocol.MsgService;
+import com.ericsson.eiffel.remrem.publish.constants.RemremPublishServiceConstants;
 import com.ericsson.eiffel.remrem.publish.helper.PublishUtils;
 import com.ericsson.eiffel.remrem.publish.helper.RMQHelper;
+import com.ericsson.eiffel.remrem.publish.service.EventTemplateHandler;
 import com.ericsson.eiffel.remrem.publish.service.MessageService;
 import com.ericsson.eiffel.remrem.publish.service.SendResult;
 import com.ericsson.eiffel.remrem.shared.VersionService;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 
 import ch.qos.logback.classic.Logger;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 
 @RestController
 @RequestMapping("/*")
+@Api(value = "REMReM Publish Service", description = "REST API for publishing Eiffel messages to message bus")
 public class ProducerController {
-
-    @Value("${generate.server.host}")
-    private String generateServerHost;
-
-    @Value("${generate.server.port}")
-    private String generateServerPort;
 
     @Autowired
     private MsgService msgServices[];
+
+    @Autowired
+    @Qualifier("messageServiceRMQImpl")
+    private MessageService messageService;
+
+    @Autowired
+    private RMQHelper rmqHelper;
+
+    @Autowired
+    private GenerateURLTemplate generateURLTemplate;
+
+    private RestTemplate restTemplate = new RestTemplate();
+
+    private JsonParser parser = new JsonParser();
+
+    private Logger log = (Logger) LoggerFactory.getLogger(ProducerController.class);
 
     public void setMsgServices(MsgService[] msgServices) {
         this.msgServices = msgServices;
     }
 
-    @Autowired
-    @Qualifier("messageServiceRMQImpl")
-    MessageService messageService;
-
-    @Autowired
-    RMQHelper rmqHelper;
-
-    RestTemplate restTemplate = new RestTemplate();
-    
     public void setRestTemplate(RestTemplate restTemplate) {
         this.restTemplate = restTemplate;
     }
 
-    Logger log = (Logger) LoggerFactory.getLogger(ProducerController.class);
-
     @SuppressWarnings({ "rawtypes", "unchecked" })
+    @ApiOperation(value = "To publish eiffel event to message bus", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Event sent successfully"),
+            @ApiResponse(code = 400, message = "Invalid event content"),
+            @ApiResponse(code = 404, message = "RabbitMq properties not found"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Service Unavailable") })
     @RequestMapping(value = "/producer/msg", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity send(@RequestParam(value = "mp", required = false) String msgProtocol,
-            @RequestParam(value = "ud", required = false) String userDomain,
-            @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "rk", required = false) String routingKey, @RequestBody JsonElement body) {
+    public ResponseEntity send(@ApiParam(value = "message protocol", required = true) @RequestParam(value = "mp") final String msgProtocol,
+                               @ApiParam(value = "user domain") @RequestParam(value = "ud", required = false) final String userDomain,
+                               @ApiParam(value = "tag") @RequestParam(value = "tag", required = false) final String tag,
+                               @ApiParam(value = "routing key") @RequestParam(value = "rk", required = false) final String routingKey,
+                               @ApiParam(value = "eiffel event", required = true) @RequestBody final JsonElement body) {
         MsgService msgService = PublishUtils.getMessageService(msgProtocol, msgServices);
 
         log.debug("mp: " + msgProtocol);
@@ -99,68 +113,91 @@ public class ProducerController {
     /**
      * This controller provides single RemRem REST API End Point for both RemRem
      * Generate and Publish.
-     * 
-     * @param mp
+     *
+     * @param msgProtocol
      *            message protocol (required)
      * @param msgType
      *            message type (required)
-     * @param ud
+     * @param userDomain
      *            user domain (not required)
      * @param tag
      *            (not required)
-     * @param rk
+     * @param routingKey
      *            (not required)
+     * @param parseData
+     *            (not required, default=false)
      * @return A response entity which contains http status and result
-     * @exception IOException
-     *                On input error.
-     * @see IOException
+     *
      * @use A typical CURL command: curl -H "Content-Type: application/json" -X POST
      *      --data "@inputGenerate_activity_finished.txt"
      *      "http://localhost:8986/generateAndPublish/?mp=eiffelsemantics&msgType=EiffelActivityFinished"
      */
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
+    @ApiOperation(value = "To generate and publish eiffel event to message bus", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Event sent successfully"),
+            @ApiResponse(code = 400, message = "Invalid event content"),
+            @ApiResponse(code = 404, message = "RabbitMq properties not found"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Message protocol is invalid") })
     @RequestMapping(value = "/generateAndPublish", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity generateAndPublish(@RequestParam(value = "mp") String mp, @RequestParam("msgType") String msgType,
-            @RequestParam(value = "ud", required = false) String userDomain,
-            @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "rk", required = false) String routingKey, @RequestBody JsonObject bodyJson) {
+    public ResponseEntity generateAndPublish(@ApiParam(value = "message protocol", required = true) @RequestParam(value = "mp") final String msgProtocol,
+                                             @ApiParam(value = "message type", required = true) @RequestParam("msgType") final String msgType,
+                                             @ApiParam(value = "user domain") @RequestParam(value = "ud", required = false) final String userDomain,
+                                             @ApiParam(value = "tag") @RequestParam(value = "tag", required = false) final String tag,
+                                             @ApiParam(value = "routing key") @RequestParam(value = "rk", required = false) final String routingKey,
+                                             @ApiParam(value = "parse data") @RequestParam(value = "parseData", required = false, defaultValue = "false") final Boolean parseData,
+                                             @ApiParam(value = "JSON message", required = true) @RequestBody final JsonObject bodyJson) {
 
-        URLTemplate urlTemplate = new URLTemplate();
-        urlTemplate.generate(mp, msgType, userDomain, routingKey, tag, generateServerHost ,generateServerPort);
-
-        ResponseEntity<String> response = null;
+        String bodyJsonOut = null;
+        if(parseData) {
+            // -- parse params in incoming request -> body -------------
+            EventTemplateHandler eventTemplateHandler = new EventTemplateHandler();
+            JsonNode parsedTemplate = eventTemplateHandler.eventTemplateParser(bodyJson.toString(), msgType);
+            bodyJsonOut = parsedTemplate.toString();
+            log.info("Parsed template: " + bodyJsonOut);
+        }else{
+            bodyJsonOut = bodyJson.toString();
+        }
 
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
-        HttpEntity<String> entity = new HttpEntity<String>(bodyJson.toString(), headers);
+        HttpEntity<String> entity = new HttpEntity<>(bodyJsonOut, headers);
 
-        String postURL = urlTemplate.getUrl();
-        Map<String, String> map = urlTemplate.getMap();
-        response = restTemplate.postForEntity(postURL, entity, String.class, map);
-        int res = response.getStatusCode().value();
+        try {
+            ResponseEntity<String> response = restTemplate.postForEntity(generateURLTemplate.getUrl(),
+                    entity, String.class, generateURLTemplate.getMap(msgProtocol, msgType));
 
-        if (res == HttpStatus.OK.value()) {
-            log.info("The result from remrem-generate is : " + res);
+            if(response.getStatusCode() == HttpStatus.OK) {
+                log.info("The result from REMReM Generate is: " + response.getStatusCodeValue());
 
-            String responseBody = response.getBody();
-            // publishing requires an array if you want status code
-            responseBody = "[" + responseBody + "]";
-            MsgService msgService = PublishUtils.getMessageService(mp, msgServices);
+                // publishing requires an array if you want status code
+                String responseBody = "[" + response.getBody() + "]";
+                MsgService msgService = PublishUtils.getMessageService(msgProtocol, msgServices);
 
-            log.debug("mp: " + mp);
-            log.debug("body: " + responseBody);
-            log.debug("user domain suffix: " + userDomain + " tag: " + tag + " Routing Key: " + routingKey);
-            if (msgService != null && mp != null) {
-                rmqHelper.rabbitMqPropertiesInit(mp);
+                log.debug("mp: " + msgProtocol);
+                log.debug("body: " + responseBody);
+                log.debug("user domain suffix: " + userDomain + " tag: " + tag + " routing key: " + routingKey);
+                if (msgService != null && msgProtocol != null) {
+                    rmqHelper.rabbitMqPropertiesInit(msgProtocol);
+                }
+                SendResult result = messageService.send(responseBody, msgService, userDomain, tag, routingKey);
+                return new ResponseEntity(result, messageService.getHttpStatus());
+            } else {
+                return response;
             }
-            SendResult result = messageService.send(responseBody, msgService, userDomain, tag, routingKey);
-            return new ResponseEntity(result, messageService.getHttpStatus());
-
-        } else {
-            log.info("The result from remrem-generate is not OK and have value: " + res);
-            return response;
+        } catch (Exception e) {
+            log.info("The result from REMReM Generate is not OK and have value: " + e.getMessage());
+            if (e.getMessage().startsWith(Integer.toString(HttpStatus.BAD_REQUEST.value()))) {
+                return new ResponseEntity(parser.parse(RemremPublishServiceConstants.GENERATE_BAD_REQUEST), HttpStatus.BAD_REQUEST);
+            } else if (e.getMessage().startsWith(Integer.toString(HttpStatus.SERVICE_UNAVAILABLE.value()))) {
+                return new ResponseEntity(parser.parse(RemremPublishServiceConstants.GENERATE_NO_SERVICE_ERROR), HttpStatus.SERVICE_UNAVAILABLE);
+            } else if (e.getMessage().startsWith(Integer.toString(HttpStatus.UNAUTHORIZED.value()))) {
+                return new ResponseEntity(parser.parse(RemremPublishServiceConstants.GENERATE_UNAUTHORIZED), HttpStatus.UNAUTHORIZED);
+            } else {
+                return new ResponseEntity(parser.parse(RemremPublishServiceConstants.GENERATE_INTERNAL_ERROR), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         }
     }
 
@@ -168,8 +205,7 @@ public class ProducerController {
      * @return this method returns the current version of publish and all loaded
      *         protocols.
      */
-
-
+    @ApiOperation(value = "To get versions of publish and all loaded protocols", response = String.class)
     @RequestMapping(value = "/versions", method = RequestMethod.GET)
     public JsonElement getVersions() {
         JsonParser parser = new JsonParser();